requirements.txt
discord.py==2.3.2
langdetect==1.0.9
python-dotenv==1.0.0 